+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * The Doc Browser will now not change the menu when refreshing.
 * Fixes an issue where URLs in the doc browser would display JSON.
 * Fixes an issue where table columns in the doc browser would be overlapping.
+* Refactor Praxis Mapper plugin to be more generic.
 
 ## 0.11.1